Change memUsageKB type from size_t to unsigned

+++ include/common.h
@@ -258,7 +258,7 @@ struct DebugInfo {
     int      activeRooms;
     int      totalEntities;
     int      aggroEntities;
-    size_t   memUsageKB;
+    unsigned memUsageKB;
     int      level;
     uint32_t seed;
     double   playerX;
